BUG #14270: Fixed external plugins.
When building external plugins that added new modules, we were getting errors
when importing modules defined by ParaView. This was due to a bugs in
pv_setup_module_environment() which resulted in changing the location where the
*.cmake files for the imported modules were searched. Updated
pv_setup_module_environment() to fix that issue and use the
vtkExternalModuleMacros file provided by VTK to simplify the logic.

Change-Id: I8d45c51eea53388bb406f445f2a493185388d894